Moved properties that were previously cluttering the MainWindow.xaml.cs file to a view model
Moved debug viewmodel button into settings

More code cleanup/ commenting
