+++ Ethenyl/SoftServe/SettingsWindow.xaml
@@ -18,6 +18,7 @@
             <RowDefinition Height="Auto"></RowDefinition>
             <RowDefinition Height="Auto"></RowDefinition>
             <RowDefinition Height="Auto"></RowDefinition>
+            <RowDefinition Height="Auto"></RowDefinition>
         </Grid.RowDefinitions>
 
         <TextBlock Grid.ColumnSpan="2" Text="{x:Static properties:Resources.SettingsHeader}" Style="{StaticResource SettingsHeader}"></TextBlock>
@@ -33,5 +34,7 @@
 
         <TextBlock Grid.Row="4" Grid.Column="0" Text="{x:Static properties:Resources.IPAddress}" ToolTip="{x:Static properties:Resources.IPAddressHelp}" Style="{StaticResource SettingsItem}"></TextBlock>
         <ListView Grid.Row="4" Grid.Column="1" ItemsSource="{Binding IPAddresses}" Style="{StaticResource SettingsSetting}"></ListView>
+        
+        <Button Grid.Row="5" Grid.ColumnSpan="2" Content="DEBUG: Get Settings ViewModel" Click="ViewModelGet_Click"></Button>
     </Grid>
 </Window>
