+++ Ethenyl/SoftServe/SettingsWindow.xaml.cs
@@ -11,6 +11,7 @@ using System.Windows.Input;
 using System.Windows.Media;
 using System.Windows.Media.Imaging;
 using System.Windows.Shapes;
+using Newtonsoft.Json;
 using SoftServe.ViewModels;
 
 namespace SoftServe
@@ -30,6 +31,16 @@ namespace SoftServe
             InitializeComponent();
             viewModel = App.ViewModel;
             this.DataContext = viewModel;
+        }
+
+        /// <summary>
+        /// Displays JSON representation of view model for easy debugging.
+        /// </summary>
+        /// <param name="sender"></param>
+        /// <param name="e"></param>
+        private void ViewModelGet_Click(object sender, RoutedEventArgs e)
+        {
+            MessageBox.Show(JsonConvert.SerializeObject(App.ViewModel, Formatting.Indented), "Current ViewModel Information");
         }
     }
 }
