+++ Ethenyl/SoftServe/SoftServe.csproj
@@ -102,6 +102,7 @@
       <DependentUpon>SettingsWindow.xaml</DependentUpon>
     </Compile>
     <Compile Include="ViewModels\EthenylViewModel.cs" />
+    <Compile Include="ViewModels\PlayerViewModel.cs" />
     <Page Include="MainWindow.xaml">
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
